+++ CHANGELOG.md
@@ -112,3 +112,4 @@ integrals to their own module
 ## v0.4.0
 * Add macros for length checking on vectors
 * Add distance method for comparing positions of two `Vec3`
+* Added interaction list octree functionality: tree build, moment updates, and traversal